requirements.txt
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.20.0
PyYAML==6.0
rsa==4.8
Telethon==1.24.0
